chore(documentation): move codeOrSourceMdx and bombArgs to utils folder
