+++ packages/documentation/src/stories/components/card/card.snapshot.stories.ts
@@ -1,7 +1,7 @@
 import type { Args, StoryContext, StoryObj } from '@storybook/web-components';
 import meta, { Default, CustomContent, CardGroup } from './card.stories';
 import { html } from 'lit';
-import { bombArgs } from '../../../utils/bombArgs';
+import { bombArgs } from '../../../utils';
 
 export default {
   ...meta,
