+++ packages/documentation/src/stories/components/collapsible/collapsible.snapshot.stories.ts
@@ -1,6 +1,6 @@
 import { html } from 'lit';
 import type { Args, StoryContext, StoryObj } from '@storybook/web-components';
-import { bombArgs } from '../../../utils/bombArgs';
+import { bombArgs } from '../../../utils';
 
 import meta, { Default } from './collapsible.stories';
