+++ packages/documentation/src/stories/misc/changelog/changelog.docs.mdx
@@ -1,6 +1,6 @@
 import { Markdown, Meta } from '@storybook/blocks';
 import changelog from '../../../../../styles/CHANGELOG.md?raw';
-import { CodeOrSourceMdx } from '../../utilities/markdown/CodeOrSourceMdx';
+import { CodeOrSourceMdx } from '../../../utils/CodeOrSourceMdx';
 
 <Meta
   title="Misc/Changelog"
