+++ packages/documentation/src/utils/index.ts
@@ -1,3 +1,4 @@
+export * from './bomb-args';
 export * from './component-properties';
 export * from './get-attributes';
 export * from './map-classes';
